Update Piece interface, changing the parameter's type of the method setValue to float type

+++ src/main/java/it/unibo/xiangqi/model/api/Piece.java
@@ -80,5 +80,5 @@ public interface Piece {
     *
     * @param value the new strategic value
     */
-    public void setValue(final int new_value);
+    public void setValue(final float new_value);
 }
